закончил CatalogScreen (убрал недоразумение в отладке)

+++ app/src/main/java/com/example/pizza_shift_2025/data/model/SizeModel.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class SizeModel(
     val name: SizeTypeModel,
-    val price: Double
+    val price: Int
 )
